Consolidate User class method tests
Tests for all user class methods in User.test.ts

Next to write tests for entry creation.
